+++ frontend/src/components/Function.tsx
@@ -10,7 +10,7 @@ export const Function = ({
   function: DocNodeFunction;
 }) => {
   return (
-    <Page>
+    <Page mode="multipage">
       <div className="p-8 pt-4">
         <div className="pb-4">
           <div className="text-gray-900 text-3xl font-medium">
