+++ frontend/src/components/TypeAlias.tsx
@@ -9,7 +9,7 @@ export const TypeAlias = ({
   typealias: DocNodeTypeAlias;
 }) => {
   return (
-    <Page>
+    <Page mode="multipage">
       <div className="p-8 pt-4">
         <div className="pb-4">
           <div className="text-gray-900 text-3xl font-medium">
