+++ frontend/src/components/Variable.tsx
@@ -6,7 +6,7 @@ import { JSDoc } from "./JSDoc";
 
 export const Variable = ({ variable }: { variable: DocNodeVariable }) => {
   return (
-    <Page>
+    <Page mode="multipage">
       <div className="p-8 pt-4">
         <div className="pb-4">
           <div className="text-gray-900 text-3xl font-medium">
